convert vote controller
