+++ app/controllers/vote.ts
@@ -1,0 +1,6 @@
+import Controller from '@ember/controller';
+import type { ModelType } from 'codecrafters-frontend/routes/vote';
+
+export default class VoteController extends Controller {
+  declare model: ModelType;
+}
