+++ app/templates/vote.hbs
@@ -1,6 +1,7 @@
 {{page-title "Vote"}}
 
 <div class="container mx-auto lg:max-w-screen-lg pt-6 md:pt-10 pb-10 md:pb-48 px-3 md:px-6">
+  {{! @glint-expect-error not ts-ified yet !}}
   <VotePage::Header />
 
   {{outlet}}
